requirements.txt
yfinance>=0.2.0
anthropic>=0.18.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
QuantLib>=1.30
